Localize domain numeric coercion
